#477158 ゲーム画面で具材となる長方形をランダムな位置から降下させる
GameLogicクラスを作り、HamburgerGameクラスから役割を分離しました。
